// front_end/contracts/Cw20Escrow.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@0.24.0.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export interface InstantiateMsg {}
export type ExecuteMsg =
  | {
      create: CreateMsg
    }
  | {
      top_up: {
        id: string
      }
    }
  | {
      set_receipient: {
        id: string
        recipient: string
      }
    }
  | {
      approve: {
        id: string
      }
    }
  | {
      refund: {
        id: string
      }
    }
  | {
      receive: Cw20ReceiveMsg
    }
export type Uint128 = string
export type Binary = string
export interface CreateMsg {
  arbiter: string
  cw20_whitelist?: string[] | null
  description: string
  end_height?: number | null
  id: string
  recipient?: string | null
  title: string
}
export interface Cw20ReceiveMsg {
  amount: Uint128
  msg: Binary
  sender: string
}
export type QueryMsg =
  | {
      list: {}
    }
  | {
      detail: {
        id: string
      }
    }
export interface DetailResponse {
  arbiter: string
  cw20_balance: Cw20Coin[]
  cw20_whitelist: string[]
  description: string
  end_height?: number | null
  id: string
  native_balance: Coin[]
  recipient?: string | null
  source: string
  title: string
}
export interface Cw20Coin {
  address: string
  amount: Uint128
}
export interface Coin {
  amount: Uint128
  denom: string
  [k: string]: unknown
}
export interface ListResponse {
  ids: string[]
}